feat: make SSH key filename optional in MikroTik SSH client
The `key_filename` parameter in the `MikroTikSSHClient` constructor is now optional, allowing users to omit it if not needed.
This change is reflected in the `MikrotikConfig` class, where the default value for `key_filename` is set to `None`.
The `DEFAULT_MIKROTIK_KEY_FILENAME` constant has been removed as it is no longer required.